Refactor UI class into Processor class
Ui class is no longer used due to switch from console outputs to using a GUI.

This results in a collection of dead functions.

Removing the unused functions and and renaming Ui class to Processor makes the code more readable for other programmers to modify in the future.

Remove the unused functions and renamed Ui to Processor. Move all input processing logic from Command into Processor.

This allows for all the relevant code to be grouped together, allowing for easier modification in the future.
